style(intl): rename pakcage and wrap long line for checkstyle
